+++ api/testsrc/no/mnemonic/services/grafeo/api/request/v1/SearchMetaFactsRequestTest.java
@@ -26,8 +26,6 @@ public class SearchMetaFactsRequestTest extends AbstractRequestTest {
             "maximum : 0.2," +
             "dimension : 'trust'," +
             "includeRetracted : true," +
-            "before : '2016-11-30T15:47:00Z'," +
-            "after : '2016-11-30T15:47:01Z'," +
             "limit : 25" +
             "}", fact);
 
@@ -42,8 +40,6 @@ public class SearchMetaFactsRequestTest extends AbstractRequestTest {
     assertEquals(0.2f, request.getMaximum(), 0.0);
     assertEquals(Dimension.trust, request.getDimension());
     assertTrue(request.getIncludeRetracted());
-    assertEquals(1480520820000L, request.getBefore().longValue());
-    assertEquals(1480520821000L, request.getAfter().longValue());
     assertEquals(25, request.getLimit().intValue());
   }